CDHSH-59 Refactored the SignUp form into several separate components, in anticipation of creating the Login scene.
