+++ services/frontend/src/components/AuthBranding/index.js
@@ -1,0 +1,1 @@
+export {default} from "./AuthBranding";
