+++ services/frontend/src/components/AuthForm/index.js
@@ -1,0 +1,1 @@
+export {default} from "./AuthForm";
