+++ services/frontend/src/components/EmailInput/EmailInput.scss
@@ -1,0 +1,6 @@
+@import "styles/_colors.scss";
+@import "styles/_dimens.scss";
+@import "styles/_mixins.scss";
+
+.email-input {
+}
