+++ services/frontend/src/components/EmailInput/index.js
@@ -1,0 +1,1 @@
+export {default} from "./EmailInput";
