+++ services/frontend/src/components/FormError/index.js
@@ -1,0 +1,1 @@
+export {default} from "./FormError";
