+++ services/frontend/src/components/LoadingButton/LoadingButton.scss
@@ -1,0 +1,12 @@
+@import "styles/_colors.scss";
+@import "styles/_dimens.scss";
+
+.loading-button {
+    .loading-button-indicator {
+        color: $color-light;
+
+        /* Need to override Material-UI's inline styles */
+        height: 24px !important;  
+        width: 24px !important;
+    }
+}
