+++ services/frontend/src/components/LoadingButton/index.js
@@ -1,0 +1,1 @@
+export {default} from "./LoadingButton";
