+++ services/frontend/src/components/index.js
@@ -1,1 +1,6 @@
+export {default as AuthBranding} from "./AuthBranding";
+export {default as AuthForm} from "./AuthForm";
+export {default as EmailInput} from "./EmailInput";
+export {default as FormError} from "./FormError";
+export {default as LoadingButton} from "./LoadingButton";
 export {default as Navbar} from "./Navbar";
