+++ services/frontend/src/scenes/SignUp/SignUp.js
@@ -1,29 +1,14 @@
-import React, {useCallback} from "react";
-import {useInput, useOnEnterKey} from "utils/hooks";
+import React from "react";
 import connect from "./connect";
 import SignUpLayout from "./SignUpLayout";
 
-const SignUp = ({errorMessage, isEmailInvalid, isLoading, onSignUp}) => {
-    const emailInput = useInput();
-    const passwordInput = useInput();
-
-    const {value: email} = emailInput;
-    const {value: password} = passwordInput;
-
-    const onSignUpClick = useCallback(() => onSignUp(email, password), [email, password, onSignUp]);
-    const onInputEnter = useCallback(useOnEnterKey(onSignUpClick), [onSignUpClick]);
-
-    return (
-        <SignUpLayout
-            emailInput={emailInput}
-            passwordInput={passwordInput}
-            isLoading={isLoading}
-            errorMessage={errorMessage}
-            isEmailInvalid={isEmailInvalid}
-            onSignUpClick={onSignUpClick}
-            onInputEnter={onInputEnter}
-        />
-    );
-};
+const SignUp = ({errorMessage, isEmailInvalid, isLoading, onSignUp}) => (
+    <SignUpLayout
+        errorMessage={errorMessage}
+        isLoading={isLoading}
+        isEmailInvalid={isEmailInvalid}
+        onSignUp={onSignUp}
+    />
+);
 
 export default connect(SignUp);
